# app/backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Avitosha API
  version: 2.0.0
  description: Gamification layer that turns useful Avito actions into Avitosha progress.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/v1/listings/{listing_id}/favorite:
    parameters: [ { $ref: '#/components/parameters/ListingID' } ]
    put:
      tags: [marketplace]
      operationId: addFavorite
      summary: Add another user's public listing to favorites
      description: "eventId is idempotent. A newly created favorite generates AD_FAVORITED server-side."
      security: [ { bearerAuth: [] }, { userIdHeader: [] } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/EventRequest' } } } }
      responses: { '200': { description: Favorite state and game feedback, content: { application/json: { schema: { $ref: '#/components/schemas/MarketplaceActionResponse' } } } } }
    delete:
      tags: [marketplace]
      operationId: removeFavorite
      summary: Remove listing from favorites
      security: [ { bearerAuth: [] }, { userIdHeader: [] } ]
      responses: { '200': { description: Favorite state, content: { application/json: { schema: { $ref: '#/components/schemas/FavoriteMutation' } } } } }
  /api/v1/me/favorites:
    get:
      tags: [marketplace]
      operationId: listFavorites
      summary: List authenticated user's favorites
      security: [ { bearerAuth: [] }, { userIdHeader: [] } ]
      parameters: [ { $ref: '#/components/parameters/PageLimit' }, { $ref: '#/components/parameters/PageOffset' } ]
      responses: { '200': { description: Favorite listings, content: { application/json: { schema: { $ref: '#/components/schemas/ListingPage' } } } } }
...
  /api/v1/listings/{listing_id}/messages:
    parameters: [ { $ref: '#/components/parameters/ListingID' } ]
    get:
      tags: [marketplace]
      operationId: listListingMessages
      summary: List messages visible to the authenticated user
      description: Returns an empty list when the user has not started a conversation. Existing messages remain visible only to their participants.
      security: [ { bearerAuth: [] }, { userIdHeader: [] } ]
      responses: { '200': { description: Messages, content: { application/json: { schema: { $ref: '#/components/schemas/ListingMessagesResponse' } } } } }
    post:
      tags: [marketplace]
      operationId: contactSeller
      summary: Send the first message to a listing seller
      description: Repeating the action returns the first stored message. The first contact generates MESSAGE_SENT server-side.
      security: [ { bearerAuth: [] }, { userIdHeader: [] } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/ContactSellerRequest' } } } }
      responses: { '201': { description: First-contact result and game feedback, content: { application/json: { schema: { $ref: '#/components/schemas/MarketplaceActionResponse' } } } } }
...
components:
  parameters:
    ListingID:
      name: listing_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    PageLimit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    PageOffset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    userIdHeader:
      type: apiKey
      in: header
      name: X-User-ID
      description: Temporary mock integration boundary; the UUID must belong to an existing user.
...
  schemas:
...
    ListingStatus:
      type: string
      enum: [DRAFT, PUBLISHED, UNPUBLISHED, SOLD]
    ListingQuality:
      type: object
      required: [score, isEligible, missingFields, nextActionHint]
      properties:
        score: { type: integer, minimum: 0, maximum: 3 }
        isEligible: { type: boolean, description: "True when the listing can be published (requires a price greater than zero)." }
        missingFields: { type: array, items: { type: string, enum: [price, photo, description] } }
        nextActionHint: { type: string }
    Listing:
      type: object
      required: [id, ownerId, categoryCode, title, description, priceKopecks, status, isDemo, photoUrls, quality]
      properties:
        id: { type: string, format: uuid }
        ownerId: { type: string, format: uuid }
        categoryCode: { type: string }
        title: { type: string, maxLength: 120 }
        description: { type: string, maxLength: 5000 }
        priceKopecks: { type: integer, format: int64, minimum: 0, example: 249000 }
        status: { $ref: '#/components/schemas/ListingStatus' }
        isDemo: { type: boolean }
        publishedAt: { type: string, format: date-time, nullable: true }
        soldAt: { type: string, format: date-time, nullable: true }
        photoUrls: { type: array, maxItems: 10, items: { type: string, format: uri } }
        quality: { $ref: '#/components/schemas/ListingQuality' }
...
    ListingPage:
      type: object
      required: [items, total, limit, offset]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Listing' } }
        total: { type: integer, minimum: 0 }
        limit: { type: integer }
        offset: { type: integer }
    FavoriteMutation:
      type: object
      required: [favorite]
      properties:
        favorite: { type: boolean }
        created: { type: boolean }
        removed: { type: boolean }
...
    ContactSellerRequest:
      type: object
      additionalProperties: false
      required: [body, eventId]
      properties:
        body: { type: string, minLength: 1, maxLength: 2000 }
        eventId: { type: string, format: uuid }
    ListingMessage:
      type: object
      required: [id, listingId, senderId, recipientId, body, createdAt]
      properties:
        id: { type: string, format: uuid }
        listingId: { type: string, format: uuid }
        senderId: { type: string, format: uuid }
        recipientId: { type: string, format: uuid }
        body: { type: string }
        createdAt: { type: string, format: date-time }
...
    ListingMessagesResponse:
      type: object
      required: [messages]
      properties:
        messages: { type: array, items: { $ref: '#/components/schemas/ListingMessage' } }
...
    EventRequest:
      type: object
      additionalProperties: false
      required: [eventId]
      properties:
        eventId: { type: string, format: uuid }
    MarketplaceActionResponse:
      type: object
      properties:
        listing: { $ref: '#/components/schemas/Listing' }
        deal: { $ref: '#/components/schemas/ListingDeal' }
        message: { $ref: '#/components/schemas/ListingMessage' }
        favorite: { type: boolean }
        counted: { type: boolean }
        first: { type: boolean }
        actionResult: { $ref: '#/components/schemas/ActionResult' }
    ListingDeal:
      type: object
      required: [id, listingId, buyerId, sellerId, deliveryUsed, completedAt]
      properties:
        id: { type: string, format: uuid }
        listingId: { type: string, format: uuid }
        buyerId: { type: string, format: uuid }
        sellerId: { type: string, format: uuid }
        deliveryUsed: { type: boolean }
        completedAt: { type: string, format: date-time }
...
    DomainEvent:
      type: object
      required: [id, type, occurredAt]
      additionalProperties: true
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
        occurredAt:
          type: string
          format: date-time
    ActionResult:
      type: object
      required: [actionId, duplicate, events]
      properties:
        actionId:
          type: string
          format: uuid
        duplicate:
          type: boolean
        events:
          type: array
          items:
            $ref: '#/components/schemas/DomainEvent'
...